.shp 를 읽는 방식에서 .zip 파일을 읽게 변경중

+++ src/main/java/com/transit/mapbox/Api/vo/ShpVo.java
@@ -1,11 +1,10 @@
 package com.transit.mapbox.Api.vo;
 
 import lombok.Data;
-import jdk.jfr.DataAmount;
 
 @Data
 public class ShpVo {
-    private int shpid;
+    private Long shpid;
     private String shpname;
     private String uploaddate;
 }
